Restructured apps.
- Rename momo to main app.
- Moved apps to dedicated apps/ directory.
