+++ apps/main/admin.py
@@ -1,5 +1,5 @@
 from django.contrib import admin
-from momo.models import UserProfile
+from .models import UserProfile
 
 # Register your models here.
 admin.site.register(UserProfile)
