+++ apps/main/forms.py
@@ -1,5 +1,5 @@
 from django import forms
-from momo.models import UserProfile
+from .models import UserProfile
 from django.contrib.auth.models import User
 
 class UserForm(forms.ModelForm):
