+++ project/urls.py
@@ -18,6 +18,6 @@ from django.urls import path, include
 
 
 urlpatterns = [
-    path('', include('momo.urls')),
+    path('', include('apps.main.urls')),
     path('admin/', admin.site.urls),
 ]
